move some methods that are only used in ConnectionManager now
